calibration: move applet to feature/touch
