+++ Marlin/src/feature/touch/ili9328.cpp
@@ -17,7 +17,7 @@
  *
  */
 
-#include "../../../inc/MarlinConfig.h"
+#include "../../inc/MarlinConfig.h"
 
 #if ENABLED(TOUCH_CALIBRATION)
