+++ Marlin/src/feature/touch/ili9341.cpp
@@ -17,144 +17,119 @@
  *
  */
 
-#include "../../../inc/MarlinConfig.h"
+#include "../../inc/MarlinConfig.h"
 
 #if ENABLED(TOUCH_CALIBRATION)
 
-#include "st7789v.h"
+#include "ili9341.h"
 #include "fsmc.h"
 
 #define LCD_Delay(ms) delay(ms)
 
-void st7789v_Init() {
+void ILI9341_Init() {
   TOUCH_LCD_IO_Init();
 
   /* Sleep In Command */
-  LCD_IO_WriteReg(ST7789V_SLEEP_IN);
-  /* Wait for 10ms */
-  LCD_Delay(10);
-
+  LCD_IO_WriteReg(ILI9341_SLEEP_IN);
   /* SW Reset Command */
-  LCD_IO_WriteReg(ST7789V_SWRESET);
+  LCD_IO_WriteReg(ILI9341_SWRESET);
   /* Wait for 200ms */
   LCD_Delay(200);
 
-  /* Sleep Out Command */
-  LCD_IO_WriteReg(ST7789V_SLEEP_OUT);
-  /* Wait for 120ms */
-  LCD_Delay(120);
-
   /* Normal display for Driver Down side */
-  LCD_IO_WriteReg(ST7789V_NORMAL_DISPLAY);
-  LCD_IO_WriteData(0xA0);
-  //LCD_IO_WriteData(0xA0);
+  LCD_IO_WriteReg(ILI9341_NORMAL_DISPLAY);
+  LCD_IO_WriteData(0xE8); // MY and ML flipped +  bit 3 RGB and BGR changed.
 
   /* Color mode 16bits/pixel */
-  LCD_IO_WriteReg(ST7789V_COLOR_MODE);
+  LCD_IO_WriteReg(ILI9341_COLOR_MODE);
   LCD_IO_WriteData(0x55);
 
   /* Set Column address CASET */
-  LCD_IO_WriteReg(ST7789V_CASET);
+  LCD_IO_WriteReg(ILI9341_CASET);
   LCD_IO_WriteData(0x00);
   LCD_IO_WriteData(0x00);
   LCD_IO_WriteData(0x01);
   LCD_IO_WriteData(0x3F);
   /* Set Row address RASET */
-  LCD_IO_WriteReg(ST7789V_RASET);
+  LCD_IO_WriteReg(ILI9341_RASET);
   LCD_IO_WriteData(0x00);
   LCD_IO_WriteData(0x00);
   LCD_IO_WriteData(0x00);
   LCD_IO_WriteData(0xEF);
 
-  /* PORCH control setting */
-  LCD_IO_WriteReg(ST7789V_PORCH_CTRL);
-  LCD_IO_WriteData(0x0C);
-  LCD_IO_WriteData(0x0C);
-  LCD_IO_WriteData(0x00);
-  LCD_IO_WriteData(0x33);
-  LCD_IO_WriteData(0x33);
-
-  /* GATE control setting */
-  LCD_IO_WriteReg(ST7789V_GATE_CTRL);
-  LCD_IO_WriteData(0x35);
-
   /* VCOM setting */
-  LCD_IO_WriteReg(ST7789V_VCOM_SET);
-  LCD_IO_WriteData(0x1F);
-
-  /* LCM Control setting */
-  LCD_IO_WriteReg(ST7789V_LCM_CTRL);
-  LCD_IO_WriteData(0x2C);
-
-  /* VDV and VRH Command Enable */
-  LCD_IO_WriteReg(ST7789V_VDV_VRH_EN);
-  LCD_IO_WriteData(0x01);
-  LCD_IO_WriteData(0xC3);
-
-  /* VDV Set */
-  LCD_IO_WriteReg(ST7789V_VDV_SET);
-  LCD_IO_WriteData(0x20);
+  LCD_IO_WriteReg(ILI9341_VCOM_CTRL1);
+  LCD_IO_WriteData(0x3E);
+  LCD_IO_WriteData(0x28);
+  LCD_IO_WriteReg(ILI9341_VCOM_CTRL2);
+  LCD_IO_WriteData(0x86);
 
   /* Frame Rate Control in normal mode */
-  LCD_IO_WriteReg(ST7789V_FR_CTRL);
-  LCD_IO_WriteData(0x0F);
+  LCD_IO_WriteReg(ILI9341_FRMCTRL);
+  LCD_IO_WriteData(0x00);
+  LCD_IO_WriteData(0x18);
 
   /* Power Control */
-  LCD_IO_WriteReg(ST7789V_POWER_CTRL);
+  LCD_IO_WriteReg(ILI9341_POWER_CTRL1);
+  LCD_IO_WriteData(0x23);
+  LCD_IO_WriteReg(ILI9341_POWER_CTRL2);
+  LCD_IO_WriteData(0x10);
 
-  //Values to be spyed on the LA.
-  //LCD_IO_WriteData(0xA4);
-  //LCD_IO_WriteData(0xA1);
+  /* Sleep Out Command */
+  LCD_IO_WriteReg(ILI9341_SLEEP_OUT);
+  /* Wait for 120ms */
+  LCD_Delay(120);
 
   /* Display ON command */
-  st7789v_DisplayOn();
+  ILI9341_DisplayOn();
 }
 
-void st7789v_DisplayOn() {
+void ILI9341_DisplayOn() {
   /* Display ON command */
-  LCD_IO_WriteReg(ST7789V_DISPLAY_ON);
+  LCD_IO_WriteReg(ILI9341_DISPLAY_ON);
+
   /* Sleep Out command */
-  LCD_IO_WriteReg(ST7789V_SLEEP_OUT);
+  LCD_IO_WriteReg(ILI9341_SLEEP_OUT);
 }
 
-void st7789v_WriteRam() {
-  LCD_IO_WriteReg(ST7789V_WRITE_RAM);
+void ILI9341_WriteRam() {
+  LCD_IO_WriteReg(ILI9341_WRITE_RAM);
 }
 
-void st7789v_SetCursor(uint16_t Xpos, uint16_t Ypos) {
+void ILI9341_SetCursor(uint16_t Xpos, uint16_t Ypos) {
   /* CASET: Comumn Addrses Set */
-  LCD_IO_WriteReg(ST7789V_CASET);
+  LCD_IO_WriteReg(ILI9341_CASET);
   LCD_IO_WriteData((Xpos >> 8) & 0xFF);
   LCD_IO_WriteData(Xpos & 0xFF);
   LCD_IO_WriteData(0x01);
   LCD_IO_WriteData(0x3F);
 
   /* RASET: Row Addrses Set */
-  LCD_IO_WriteReg(ST7789V_RASET);
+  LCD_IO_WriteReg(ILI9341_RASET);
   LCD_IO_WriteData((Ypos >> 8) & 0xFF);
   LCD_IO_WriteData(Ypos & 0xFF);
   LCD_IO_WriteData(0x00);
   LCD_IO_WriteData(0xEF);
 
-  LCD_IO_WriteReg(ST7789V_WRITE_RAM);
+  LCD_IO_WriteReg(ILI9341_WRITE_RAM);
 }
 
-void st7789v_SetWindow(uint16_t Xmin, uint16_t Ymin, uint16_t Xmax, uint16_t Ymax) {
+void ILI9341_SetWindow(uint16_t Xmin, uint16_t Ymin, uint16_t Xmax, uint16_t Ymax) {
   /* CASET: Comumn Addrses Set */
-  LCD_IO_WriteReg(ST7789V_CASET);
+  LCD_IO_WriteReg(ILI9341_CASET);
   LCD_IO_WriteData((Xmin >> 8) & 0xFF);
   LCD_IO_WriteData(Xmin & 0xFF);
   LCD_IO_WriteData((Xmax >> 8) & 0xFF);
   LCD_IO_WriteData(Xmax & 0xFF);
 
   /* RASET: Row Addrses Set */
-  LCD_IO_WriteReg(ST7789V_RASET);
+  LCD_IO_WriteReg(ILI9341_RASET);
   LCD_IO_WriteData((Ymin >> 8) & 0xFF);
   LCD_IO_WriteData(Ymin & 0xFF);
   LCD_IO_WriteData((Ymax >> 8) & 0xFF);
   LCD_IO_WriteData(Ymax & 0xFF);
 
-  LCD_IO_WriteReg(ST7789V_WRITE_RAM);
+  LCD_IO_WriteReg(ILI9341_WRITE_RAM);
 }
 
 #endif
